# openapi.yaml
openapi: 3.0.1
info:
  title: The OpenShelf API
  version: "2.0"
  description: |
    # Summary
    The OpenShelf  API
servers:
  - url: https://TheOpenShelf.dev
    description: The OpenShelf API
paths:
...
  /hub/items:
    get:
      operationId: getItems
      summary: Retrieve all items (filter by current user items)
      tags:
        - items-hub-api
      parameters:
        - name: borrowedByCurrentUser
          in: query
          description: Filter items borrowed by current user
          required: false
          schema:
            type: boolean
        - name: borrowedBy
          in: query
          description: Filter items by borrowed by
          required: false
          schema:
            type: string
        - name: libraryIds
          in: query
          description: Filter items by library IDs
          required: false
          schema:
            type: array
            items:
              type: string
        - name: communityIds
          in: query
          description: Filter items by community IDs
          required: false
          schema:
            type: array
            items:
              type: string
        - name: categories
          in: query
          description: Filter items by categories
          required: false
          schema:
            type: array
            items:
              type: string
        - name: searchText
          in: query
          description: Filter items by search text in name or description
          required: false
          schema:
            type: string
        - name: currentlyAvailable
          in: query
          description: Filter items by availability
          required: false
          schema:
            type: boolean
        - name: sortBy
          in: query
          description: Sort items by a specific field
          required: false
          schema:
            type: string
            enum: [createdAt, borrowCount, favorite]
        - name: sortOrder
          in: query
          description: Sort order (asc or desc)
          required: false
          schema:
            type: string
            enum: [asc, desc]
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            default: 10
        - name: startDate
          in: query
          description: Filter items by start date
          required: false
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: Filter items by end date
          required: false
          schema:
            type: string
            format: date
        - name: favorite
          in: query
          description: Filter items by favorite status
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: A paginated list of items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedItemsResponse"
    post:
      operationId: addItem
      summary: Add a new item
      tags:
        - items-hub-api
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Item"
      responses:
        "201":
          description: Item created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
  /hub/items/{item-id}:
    parameters:
      - $ref: "#/components/parameters/itemId"
    get:
      operationId: getItem
      summary: Retrieve a specific item by ID
      tags:
        - items-hub-api
      responses:
        "200":
          description: A single item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
  /hub/items/{item-id}/borrow-records:
    parameters:
      - $ref: "#/components/parameters/itemId"
      - name: owner
        in: query
        description: Filter by owner
        required: false
        schema:
          type: string
      - name: borrower
        in: query
        description: Filter by borrower
        required: false
        schema:
          type: string
    get:
      operationId: getItemBorrowRecords
      summary: Retrieve borrow records for a specific item
      tags:
        - items-hub-api
      responses:
        "200":
          description: Borrow records for the item
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BorrowRecord"
    delete:
      operationId: deleteBorrowRecord
      summary: Delete a borrow record for a specific item
      tags:
        - items-hub-api
      parameters:
        - name: recordId
          in: query
          description: ID of the borrow record to delete
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Borrow record deleted successfully
  /hub/items/{item-id}/actions/borrow:
    parameters:
      - $ref: "#/components/parameters/itemId"
    post:
      operationId: borrowItem
      summary: Borrow an item
      tags:
        - items-hub-api
      requestBody:
        description: Borrow item details
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                borrowBy:
                  type: string
                  format: uuid
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
      responses:
        "201":
          description: Item borrowed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
  /hub/items/{item-id}/actions/favorite:
    parameters:
      - $ref: "#/components/parameters/itemId"
    post:
      operationId: markAsFavorite
      summary: Toggle favorite status of an item
      tags:
        - items-hub-api
      responses:
        "200":
          description: Item favorite status toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
  /hub/items/{item-id}/actions/return-item:
    parameters:
      - $ref: "#/components/parameters/itemId"
    post:
      operationId: returnItem
      summary: Return a borrowed item
      tags:
        - items-hub-api
      requestBody:
        description: Return item details
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                borrowRecordId:
                  type: string
                  format: uuid
                  example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
              required:
                - borrowRecordId
      responses:
        "200":
          description: Item returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
...
  /hub/items/{item-id}/actions/approval-reservation:
    parameters:
      - $ref: "#/components/parameters/itemId"
    post:
      operationId: approvalReservation
      summary: Approve or reject a reservation
      tags:
        - items-hub-api
      requestBody:
        description: Approval details
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                decision:
                  type: string
                  enum: [approve, reject]
                borrowRecordId:
                  type: string
                  format: uuid
                  example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
              required:
                - decision
                - borrowRecordId
      responses:
        "200":
          description: Reservation approval status updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
  /hub/items/{item-id}/actions/approval-pickup:
    parameters:
      - $ref: "#/components/parameters/itemId"
    post:
      operationId: approvalPickup
      summary: Approve or reject a pickup
      tags:
        - items-hub-api
      requestBody:
        description: Approval details
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                decision:
                  type: string
                  enum: [approve, reject]
                borrowRecordId:
                  type: string
                  format: uuid
                  example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
              required:
                - decision
                - borrowRecordId
      responses:
        "200":
          description: Pickup approval status updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
  /hub/items/{item-id}/actions/approval-return:
    parameters:
      - $ref: "#/components/parameters/itemId"
    post:
      operationId: approvalReturn
      summary: Approve or reject a return
      tags:
        - items-hub-api
      requestBody:
        description: Approval details
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                decision:
                  type: string
                  enum: [approve, reject]
                borrowRecordId:
                  type: string
                  format: uuid
                  example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
              required:
                - decision
                - borrowRecordId
      responses:
        "200":
          description: Return approval status updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
...
  /hub/libraries:
    get:
      operationId: getHubLibraries
      summary: Retrieve all hub libraries
      tags:
        - libraries-hub-api
      parameters:
        - name: communityIds
          in: query
          description: Filter libraries by community IDs
          required: false
          schema:
            type: array
            items:
              type: string
              format: uuid
      responses:
        "200":
          description: A list of hub libraries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Library"
    post:
      operationId: addHubLibrary
      summary: Add a new hub library
      tags:
        - libraries-hub-api
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Library"
      responses:
        "201":
          description: Library created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Library"
  /hub/libraries/{library-id}:
    parameters:
      - $ref: "#/components/parameters/libraryId"
    get:
      operationId: getHubLibrary
      summary: Retrieve a specific hub library by ID
      tags:
        - libraries-hub-api
      responses:
        "200":
          description: A single hub library
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Library"
    put:
      operationId: updateHubLibrary
      summary: Update an existing hub library
      tags:
        - libraries-hub-api
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Library"
      responses:
        "200":
          description: Library updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Library"
    delete:
      operationId: deleteHubLibrary
      summary: Delete a hub library
      tags:
        - libraries-hub-api
      responses:
        "204":
          description: Library deleted successfully
  /hub/libraries/{library-id}/members:
    parameters:
      - $ref: "#/components/parameters/libraryId"
    get:
      operationId: getLibraryMembers
      summary: Retrieve members of a specific hub library by ID
      tags:
        - libraries-hub-api
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: A paginated list of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedLibraryMembersResponse"
    post:
      operationId: addMember
      summary: Add a member to a specific hub library by ID
      tags:
        - libraries-hub-api
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryMember"
      responses:
        "201":
          description: Member added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryMember"
...
components:
...
  securitySchemes:
    BearerAuthOAuth:
      description: Bearer Token.
      type: http
      scheme: bearer
  parameters:
    itemId:
      name: item-id
      in: path
      description: Item ID
      required: true
      schema:
        type: string
        format: uuid
...
    libraryId:
      name: library-id
      in: path
      description: Library ID
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    User:
      type: object
      required:
        - id
        - login
        - username
        - firstName
        - lastName
        - roles
      properties:
        id:
          type: string
          format: uuid
        disabled:
          type: boolean
        isEmailVerified:
          type: boolean
        username:
          type: string
        avatarUrl:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        flatNumber:
          type: string
          example: A701
        roles:
          type: array
          items:
            type: string
            example: "admin"
        streetAddress:
          type: string
        city:
          type: string
        postalCode:
          type: string
        country:
          type: string
        preferredLanguage:
          type: string
          example: "en"
    LibraryMember:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            role:
              type: string
              enum: ["admin", "member", "requestingJoin"]
              example: "member"
    Item:
      type: object
      required:
        - id
        - name
        - owner
        - imageUrl
        - description
        - shortDescription
        - category
        - favorite
        - borrowCount
        - libraryId
        - borrowRecords

      properties:
        id:
          type: string
          format: uuid
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        name:
          type: string
          example: "Harry Potter Book"
        owner:
          type: string
          example: "TheOpenShelf"
        imageUrl:
          type: string
          example: "/items/harry-potter.png"
        description:
          type: string
          example: "A magical adventure story that follows the journey of a young wizard and his friends."
        shortDescription:
          type: string
          example: "A magical adventure story."
        category:
          $ref: "#/components/schemas/Category"
        favorite:
          type: boolean
          example: true
        borrowCount:
          type: integer
          example: 8
        libraryId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        createdAt:
          type: string
          format: date-time
          example: "2023-11-15T13:45:30Z"
        borrowRecords:
          type: array
          items:
            $ref: "#/components/schemas/BorrowRecord"
    BorrowStatus:
      type: string
      enum:
        - reserved-unconfirmed
        - reserved-confirmed
        - reserved-ready-to-pickup
        - reserved-pickup-unconfirmed
        - borrowed-active
        - borrowed-due-today
        - borrowed-late
        - borrowed-return-unconfirmed
        - returned-early
        - returned-on-time
        - returned-late
    BorrowRecord:
      type: object
      required:
        - id
        - borrowedBy
        - startDate
        - endDate
        - reservationDate
        - status
      properties:
        id:
          type: string
          format: uuid
        borrowedBy:
          type: string
          example: "someone_else@example.com"
        pickupDate:
          type: string
          format: date
          example: "2023-11-15"
        startDate:
          type: string
          format: date
          example: "2023-11-15"
        endDate:
          type: string
          format: date
          example: "2023-11-22"
        reservationDate:
          type: string
          format: date
          example: "2023-11-15"
        effectiveReturnDate:
          type: string
          format: date
          example: "2023-11-22"
        status:
          $ref: "#/components/schemas/BorrowStatus"
...
    Category:
      type: object
      required:
        - id
        - name
        - color
        - template
        - icon
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        template:
          type: string
...
    PaginatedItemsResponse:
      type: object
      required:
        - totalPages
        - totalItems
        - currentPage
        - itemsPerPage
        - items
      properties:
        totalPages:
          type: integer
          example: 10
        totalItems:
          type: integer
          example: 100
        currentPage:
          type: integer
          example: 1
        itemsPerPage:
          type: integer
          example: 10
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
...
    PaginatedLibraryMembersResponse:
      type: object
      required:
        - totalPages
        - totalItems
        - currentPage
        - usersPerPage
        - users
      properties:
        totalPages:
          type: integer
          example: 10
        totalItems:
          type: integer
          example: 100
        currentPage:
          type: integer
          example: 1
        itemsPerPage:
          type: integer
          example: 10
        items:
          type: array
          items:
            $ref: "#/components/schemas/LibraryMember"
...
    Library:
      type: object
      required:
        - id
        - name
        - requiresApproval
        - freeAccess
        - location
        - instructions
        - isAdmin
        - communityId
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        requiresApproval:
          type: boolean
        freeAccess:
          type: boolean
        location:
          $ref: "#/components/schemas/Location"
        instructions:
          type: string
        isAdmin:
          type: boolean
          example: true
        communityId:
          type: string
          format: uuid
          description: The ID of the community to which the library belongs
    Location:
      type: object
      required:
        - name
        - address
        - coordinates
      properties:
        name:
          type: string
        address:
          type: string
        coordinates:
          type: object
          properties:
            lat:
              type: number
            lng:
              type: number
...